% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siete_dias.R
\name{incidencias_siete_dias}
\alias{incidencias_siete_dias}
\title{Datos internacionales COVID-19 7 dias}
\usage{
incidencias_siete_dias(fecha_de_trabajo)
}
\description{
Datos internacionales COVID-19 7 dias
}
